feat: правки согласно замечаниям
