+++ build.gradle
@@ -8,6 +8,7 @@ group = 'com.panyukovnn'
 version = '1.0-SNAPSHOT'
 
 repositories {
+    mavenLocal()
     mavenCentral()
 }
 
@@ -20,9 +21,12 @@ dependencies {
     testAnnotationProcessor 'org.projectlombok:lombok'
 
     implementation 'org.apache.commons:commons-lang3'
+//    implementation 'com.panyukovn:logging-starter-maven:1.0-SNAPSHOT'
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testImplementation 'org.junit.jupiter:junit-jupiter'
+
+
 }
 
 test {
