--- src/main/java/com/panyukovnn/linkShortener/exceptions/NotFoundException.java
@@ -1,8 +1,0 @@
-package com.panyukovnn.linkShortener.exceptions;
-
-public class NotFoundException extends LinkShortenerException{
-
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
