--- src/main/java/com/panyukovnn/linkShortener/service/LinkInfoService.java
@@ -1,15 +1,0 @@
-package com.panyukovnn.linkShortener.service;
-
-import com.panyukovnn.linkShortener.dto.CreateShortLinkRequest;
-import com.panyukovnn.linkShortener.model.LinkInfoResponse;
-
-import java.util.List;
-
-public interface LinkInfoService {
-
-    LinkInfoResponse getByShortLink(String shortLink);
-
-    List<LinkInfoResponse> findByFilter();
-
-    LinkInfoResponse createLinkInfo(CreateShortLinkRequest request);
-}
