+++ src/main/java/com/panyukovnn/linkshortener/exceptions/NotFoundException.java
@@ -1,0 +1,8 @@
+package com.panyukovnn.linkshortener.exceptions;
+
+public class NotFoundException extends LinkShortenerException {
+
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
