+++ src/main/java/com/panyukovnn/linkshortener/util/Constants.java
@@ -1,4 +1,4 @@
-package com.panyukovnn.linkShortener.util;
+package com.panyukovnn.linkshortener.util;
 
 public class Constants {
